Changed zero on num pad layer back to space and added . and ,
The occasional zero in place of the space was very hard to keep track
of, so I moved it back to the center column.

I also moved . up to the left center column in preparation for a split
keyboard with only three thumb buttons. I added , right above it since
I often found myself switching between the main and num pad layer for
the comma when entering a list of numbers.