Combined layer keys and backspace/delete keys
Combining the two makes the layer key holding keys. This should no
longer be a problem since the modifiers are now on the home row. Having
a separate key for the backspace and delete keys was causing my thumbs
to move too much. Maybe on a split board I can have more thumb keys, but
I can only use two on this board comfortably.