Moved thumb keys in toward center
My thumb was cramping up too much when using the layer key underneath
the pointer finger keys, at least on the left side. I have to stretch my
thumb a bit to reach the middle keys, but it may be better than cramping
it up trying to use the outer key where it is now.